Include definition of Name class in last example
I had a hard time understanding this example. Part of it was due to the typo in pull request 3, but another part was because I didn't realize immediately that Person had a list of Name objects instead of Strings.

Every other example stands alone from the others, so I was surprised this one didn't.

However, this isn't a clear improvement, so I'll understand if you reject this.